requirements.txt
Flask==3.1.0
flask-cors==5.0.0
Flask-Injector==0.15.0
gunicorn==23.0.0
protobuf==5.29.3
aiohttp==3.11.12
wheel==0.44.0
click==8.1.8
MarkupSafe==3.0.2
Werkzeug==3.1.3
mkdocstrings[python]==0.28.1
mkdocs-material==9.6.4
itsdangerous==2.2.0
Jinja2==3.1.5
mkdocs==1.6.1
SQLAlchemy==2.0.38
tabulate==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
cryptography==44.0.1
pika==1.3.2
sql==0.4.0
psycopg2-binary==2.9.7
yt-dlp==2024.7.16

# Utils
transformers==4.45.2
tiktoken==0.9.0
python-dotenv==1.0.1
pydantic==2.10.6
flask-executor==1.0.0
flask-sqlalchemy==3.0.0
retry==0.9.2
XlsxWriter==3.2.2
PyMuPDF==1.23.6
pytesseract==0.3.10
openpyxl==3.1.2
networkx==3.2.1
python-pptx==0.6.23
unstructured==0.15.12
GitPython==3.1.40
beautifulsoup4==4.12.2
sentry-sdk==2.21.0
pdfplumber==0.11.0
aio-pika==9.0.5

# AI & core services
nomic==3.4.1
openai==1.63.0
langchain==0.3.18
langchainhub==0.1.21
langchain_openai==0.3.5
langchain_qdrant==0.2.0
langchain_neo4j==0.4.0
langchain_community==0.3.16
langgraph==0.2.69
canvasapi==3.2.0

# Data
boto3==1.36.20
qdrant-client==1.13.2
supabase==2.13.0
minio==7.2.15
redis[hiredis]==5.2.1

# Logging 
posthog==3.13.0
sentry-sdk==2.21.0

ollama==0.4.7

# File processing libraries for chat uploads
PyMuPDF==1.23.6
pdfplumber==0.11.0
pytesseract==0.3.10 # image OCR
python-docx==0.8.11 # DOCX files
openpyxl==3.1.2 # excel
python-pptx==0.6.23 # PowerPoint
beautifulsoup4==4.12.2 # HTML parsing
pysrt==1.1.2 # SRT subtitle files
Pillow==10.0.0 # Image processing for OCR
docx2txt==0.8 # Alternative DOCX processing
# Not currently supporting coursera ingest
# cs-dlp @ git+https://github.com/raffaem/cs-dlp.git@0.12.0b0 # previously called coursera-dl

# removed due to /ingest in Beam
# canvasapi
# GitPython
# pysrt
# docx2txt
# pydub
# ffmpeg-python
# ffprobe
# ffmpeg
# beautifulsoup4
# PyMuPDF
# pytesseract # image OCR
# openpyxl # excel
# networkx # unused part of excel partitioning :(
# python-pptx
# unstructured # causes huge ~5.3 GB of installs. Probbably from onnx: https://github.com/Unstructured-IO/unstructured/blob/ad14321016533dc03c1782f6ebea00bc9c804846/requirements/extra-pdf-image.in#L4

# pdf packages for unstructured
# pdf2image
# pdfminer.six
# opencv-python-headless
# unstructured.pytesseract
# unstructured-inference # this is the real large one :(
# unstructured[xlsx,image,pptx] # causes huge ~5.3 GB of installs. Probbably from onnx: https://github.com/Unstructured-IO/unstructured/blob/ad14321016533dc03c1782f6ebea00bc9c804846/requirements/extra-pdf-image.in#L4

azure-storage-blob
chromadb
pyodbc
sqlalchemy